refactor: 상품 삭제시 상품을 DB에서 갸져오지 말기
예전에 findById를 이용해 상품이 존재하는지 확인하던걸 existsById를
이용하는 것으로 바꾸었습니다.

+++ src/main/java/sparta/spartateamproject1/service/impl/ItemServiceImpl.java
@@ -136,9 +136,9 @@ public class ItemServiceImpl implements ItemService {
         // TODO: item 삭제를 할수있는 관리자는 누구일까요?
         getAdminIfExistsAndActive(attr.getId());
 
-        Item item = itemRepository.findById(itemId).orElseThrow(
-            () -> new ItemNotFoundException("존재하지 않는 상품 입니다.")
-        );
+        if (!itemRepository.existsById(itemId)) {
+            throw new ItemNotFoundException("존재하지 않는 상품 입니다.");
+        }
 
         itemRepository.deleteById(itemId);
     }
